+++ app/views/subject_plans/semester_card/_add_subject_form.html.erb
@@ -6,7 +6,7 @@
   data: {
     turbo: true,
     controller: "subject-selector",
-    "subject-selector-url-value" => subject_plan_subjects_path,
+    "subject-selector-url-value": planner_subjects_path,
     action: "click->subject-selector#onClick"
   },
   class: 'flex items-center gap-3 px-4 py-3 hover:bg-gray-100 border-t border-gray-100'
